changed price field to float type

+++ orders/models.py
@@ -5,7 +5,7 @@ from users.models import User
 class CartItem(models.Model):
 
     name = models.CharField(max_length=254, blank=False, verbose_name="product name")
-    price = models.IntegerField(default=0, verbose_name="product price")
+    price = models.FloatField(default=0, verbose_name="product price")
     thumbnail = models.URLField(max_length=254, blank=False, verbose_name="product thumbnail")
     quantity = models.IntegerField(default=1, verbose_name="number of items")
     created_date = models.DateTimeField(auto_now_add=True)
